HiGlass scale 1d track improvements
